Cleaned up code
- Changed some file names in the DataAccess and DataSourcess folders
- Added new file CurrentUserJson to clean up my code
- Cleaned up the UserRegister and UserLogin by changing the way the info is written to the json
- Fixed the UserRegister > CurrentUser bug (UserID is now saved correctly)
- Made some minor changes to the MainMenu file
