+++ DataAccess/UpdateUserJson.cs
@@ -4,13 +4,11 @@ namespace Restaurant
 {
     class UpdateJson
     {
-
         public UpdateJson()
         {
 
         }
-        public void writeToJson(string firstName, string lastName, string userName, string phoneNumber,
-        string passWord, string email, List<Address> address, string path)
+        public void writeToJson(UserInfo newUser, string path)
         {
             // check if json file is not empty
             if (File.Exists(path) && new FileInfo(path).Length > 0)
@@ -21,21 +19,12 @@ namespace Restaurant
                 // Deserialize the JSON into a list of User objects
                 List<UserInfo> users = JsonConvert.DeserializeObject<List<UserInfo>>(json);
 
+                // Create UserID (add 1 compared to previous user in the list)
+                int nextUserID = users.Count > 0 ? users[users.Count - 1].UserID + 1 : 1;
+                newUser.UserID = nextUserID;
 
-                // Create a list of user information
-                UserInfo newUser = new UserInfo
-                {
-                    UserID = users.Count > 0 ? users[users.Count - 1].UserID + 1 : 1,
-                    UserName = userName,
-                    PassWord = passWord,
-                    FirstName = firstName,
-                    LastName = lastName,
-                    Email = email,
-                    PhoneNumber = phoneNumber,
-                    Address = address
-                };
+                // add user to the user list
                 users.Add(newUser);
-
 
                 // Convert the list to JSON
                 json = JsonConvert.SerializeObject(users, Formatting.Indented);
@@ -47,12 +36,12 @@ namespace Restaurant
             // if json file is empty, create first account with ID number 1
             else
             {
+                newUser.UserID = 1;
+
                 // Create a list of user information
                 List<UserInfo> firstUser = new List<UserInfo>
                 {
-                    new UserInfo { UserID = 1,
-                    UserName = userName, PassWord = passWord, FirstName = firstName, LastName = lastName,
-                    Email = email, PhoneNumber = phoneNumber, Address = address}
+                    newUser
                 };
 
                 // Convert the list to JSON
@@ -61,8 +50,6 @@ namespace Restaurant
                 // Write the JSON to a file
                 File.WriteAllText(path, json);
             }
-
         }
-
     }
 }
