--- DataSources/User.json
@@ -1,70 +1,0 @@
-[
-  {
-    "UserID": 1,
-    "UserName": "Uname",
-    "PassWord": "Examplepassword1",
-    "FirstName": "Fname",
-    "LastName": "Lname",
-    "Email": "example@email.com",
-    "PhoneNumber": "012345678",
-    "Address": [
-      {
-        "City": "Example City",
-        "Street": "Example Street",
-        "HouseNumber": "Example Housenumber",
-        "ZipCode": "Example Zipcode"
-      }
-    ]
-  },
-  {
-    "UserID": 2,
-    "UserName": "newUserName",
-    "PassWord": "Password123",
-    "FirstName": "name",
-    "LastName": "lastname",
-    "Email": "newemail@email.com",
-    "PhoneNumber": "phonenumber",
-    "Address": [
-      {
-        "City": "city",
-        "Street": "street",
-        "HouseNumber": "housenumber",
-        "ZipCode": "1234AB"
-      }
-    ]
-  },
-  {
-    "UserID": 3,
-    "UserName": "Wowwie",
-    "PassWord": "SuperWow123",
-    "FirstName": "Wow",
-    "LastName": "Wie",
-    "Email": "wow@email.com",
-    "PhoneNumber": "0698765432",
-    "Address": [
-      {
-        "City": "WowCity",
-        "Street": "WowStreet",
-        "HouseNumber": "666",
-        "ZipCode": "1234AB"
-      }
-    ]
-  },
-  {
-    "UserID": 4,
-    "UserName": "myUserName",
-    "PassWord": "myPass123",
-    "FirstName": "myName",
-    "LastName": "myLastName",
-    "Email": "myemail@email.com",
-    "PhoneNumber": "0101234567",
-    "Address": [
-      {
-        "City": "myCity",
-        "Street": "myStreet",
-        "HouseNumber": "123",
-        "ZipCode": "1234AB"
-      }
-    ]
-  }
-]
